Feature be board

See merge request cloud_track/class_05/web_project1/team04/art-board!18

+++ src/main/java/com/elice/artBoard/board/service/BoardService.java
@@ -5,6 +5,7 @@ import com.elice.artBoard.board.domain.BoardImage;
 import com.elice.artBoard.board.dto.RequestBoardForm;
 import com.elice.artBoard.board.dto.ResponseBoardForm;
 import com.elice.artBoard.board.repository.BoardRepository;
+import com.elice.artBoard.member.entity.Member;
 import com.elice.artBoard.post.entity.Post;
 import com.elice.artBoard.post.service.PostService;
 import lombok.RequiredArgsConstructor;
@@ -28,26 +29,38 @@ public class BoardService {
 
 
     @Transactional
-    public Board save(RequestBoardForm form) {
-        Board board = Board.create(form.getTitle(), form.getDescription());
+    public Board save(RequestBoardForm form, Member member) {
+        Board board = Board.create(form.getTitle(), form.getDescription(), member);
         return boardRepository.save(board);
     }
 
     @Transactional
-    public Board update(Long boardId, RequestBoardForm form) {
+    public Board update(Long boardId, RequestBoardForm form, Member member) {
         Board findBoard = boardRepository.findById(boardId).orElseThrow(() -> new IllegalArgumentException("일치하는 게시판이 없습니다"));
+
+        if (!member.getMemberId().equals(findBoard.getMember().getMemberId())) {
+            throw new IllegalArgumentException("회원 정보가 일치하지 않습니다");
+        }
+
         return findBoard.update(form.getTitle(), form.getDescription());
     }
 
     @Transactional
-    public void delete(Long boardId) {
+    public void delete(Long boardId, Member member) {
+
         Board board = boardRepository.findById(boardId).orElseThrow(() -> new IllegalArgumentException("일치하는 게시판이 없습니다"));
+
+        if (!member.getMemberId().equals(board.getMember().getMemberId())) {
+            throw new IllegalArgumentException("회원 정보가 일치하지 않습니다");
+        }
+
         Post post = postService.getPostByBoardId(boardId);
 
         if (post != null) {
             postService.deletePost(post.getId());
         }
 
+        boardImageService.delete(board.getId());
         boardRepository.delete(board);
     }
 
@@ -73,7 +86,9 @@ public class BoardService {
         IntStream.range(0, bound).forEach(i -> {
             Board board = boards.get(i);
             BoardImage boardImage = images.get(i);
-            formList.add(new ResponseBoardForm(board.getId(), board.getTitle(), board.getDescription(), boardImage.getId()));
+            formList.add(new ResponseBoardForm(
+                    board.getId(), board.getTitle(), board.getDescription(), boardImage.getId(), board.getMember().getMemberId()
+            ));
         });
 
         return formList;
